add web_content, it is unfinished (for example it still needs a public_content_t equivalent)
The idea is to split web content from the various web servers. Not every web server is the same. Compare for example lighttp versus nginx

One this they all have in common though is that they serve content. By splitting the policy for the content from the various webserver policies, we can share the web content between various webserver implementations