setfscreate subj: at least it is consistent now
if setfscreatecon is attempted but denied than the panick reaction is to "get enforce mode"

you would say then: so why allow get enforce mode if setfscreatecon is allowed?

i would say: well because i figured this out late in the game and some domains use setfscreatecon and get_enforce_mode and i dont know if the get_enforce_mode was
added because setfscreatecon was denied, or if it was added for other reasons.

allowing get_enforce_mode to all setfscreateconners isnt that bad. sure its not optimal either but consider this: at least now it is consistent.